+++ jupyterlab_bigquery/src/components/details_panel/dataset_details_panel.tsx
