+++ jupyterlab_bigquery/src/components/details_panel/service/list_dataset_details.ts
