+++ jupyterlab_bigquery/src/components/details_panel/service/list_table_details.ts
