+++ jupyterlab_bigquery/src/components/details_panel/table_details_widget.tsx
