+++ jupyterlab_bigquery/src/components/list_items_panel/list_tree_item_widget.tsx
@@ -8,7 +8,7 @@ import { JupyterFrontEnd } from '@jupyterlab/application';
 
 import { ListProjectsService, Projects } from './service/list_items';
 import { ListProjectItem } from './list_tree_item';
-import { WidgetManager } from '../widget_manager';
+import { WidgetManager } from '../../utils/widget_manager';
 import { QueryEditorTabWidget } from '../query_editor/query_editor_tab/query_editor_tab_widget';
 
 interface Props {
