+++ jupyterlab_bigquery/src/components/query_editor/query_editor_tab/query_editor_tab.tsx
