+++ jupyterlab_bigquery/src/components/query_editor/query_text_editor/service/query.ts
