+++ jupyterlab_bigquery/src/index.ts
@@ -6,9 +6,9 @@ import {
   JupyterFrontEndPlugin,
 } from '@jupyterlab/application';
 
-import { ListItemsWidget } from './list_items_panel/list_tree_item_widget';
-import { ListProjectsService } from './list_items_panel/service/list_items';
-import { WidgetManager } from './widget_manager';
+import { ListItemsWidget } from './components/list_items_panel/list_tree_item_widget';
+import { ListProjectsService } from './components/list_items_panel/service/list_items';
+import { WidgetManager } from './utils/widget_manager';
 
 async function activate(app: JupyterFrontEnd) {
   const manager = new WidgetManager(app);
