+++ jupyterlab_bigquery/src/utils/widget_manager.ts
